[feat] inquiryContext 생성
